requirements.txt
alembic==1.5.4
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy=alembic==1.5.4
attrs==20.3.0
Authlib==0.14.1 # https://github.com/Vuka951/tutorial-code/blob/master/flask-google-oauth2/requirements.txt 
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.2
cryptography==2.9
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
flask-login==0.5.0 # added 8.6.21; was 'flask-login-0.5.0' in terminal after pip install flask_login was run in venv 
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.1.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1=1.3.23
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1